factor convert_closure from ode to custom_derivatives.
