+++ jax/custom_derivatives.py
@@ -19,10 +19,11 @@ import operator as op
 from typing import Callable, Sequence, Tuple, Any
 
 from . import core
+from . import dtypes
 from . import linear_util as lu
 from .tree_util import (tree_flatten, tree_unflatten, tree_map, tree_multimap,
                         register_pytree_node_class)
-from .util import safe_zip, safe_map, split_list
+from .util import cache, safe_zip, safe_map, split_list
 from .api_util import flatten_fun_nokwargs, argnums_partial, wrap_hashably
 from .core import raise_to_shaped
 from .ad_util import Zero, zeros_like_aval, stop_gradient_p
@@ -837,3 +838,52 @@ class Residuals:
   def tree_unflatten(cls, aux, consts):
     jaxpr, in_tree, out_tree = aux
     return cls(jaxpr, in_tree, out_tree, consts)
+
+
+def closure_convert(fun, example_args):
+  flat_args, in_tree = tree_flatten(example_args)
+  in_avals = tuple(map(abstractify, flat_args))
+  return _closure_convert_for_avals(fun, in_tree, in_avals)
+
+@cache()
+def _closure_convert_for_avals(fun, in_tree, in_avals):
+  if config.omnistaging_enabled:
+    wrapped_fun, out_tree = flatten_fun_nokwargs(lu.wrap_init(fun), in_tree)
+    jaxpr, out_pvals, consts = pe.trace_to_jaxpr_dynamic(wrapped_fun, in_avals)
+  else:
+    in_pvals = [pe.PartialVal.unknown(aval) for aval in in_avals]
+    wrapped_fun, out_tree = flatten_fun_nokwargs(lu.wrap_init(fun), in_tree)
+    with core.initial_style_staging():  # type: ignore
+      jaxpr, out_pvals, consts = pe.trace_to_jaxpr(
+        wrapped_fun, in_pvals, instantiate=True, stage_out=False)  # type: ignore
+  out_tree = out_tree()
+
+  # We only want to closure convert for constants with respect to which we're
+  # differentiating. As a proxy for that, we hoist consts with float dtype.
+  # TODO(mattjj): revise this approach
+  from .numpy import inexact
+  is_float = lambda c: dtypes.issubdtype(dtypes.dtype(c), inexact)
+  (closure_consts, hoisted_consts), merge = partition_list(is_float, consts)
+  num_consts = len(hoisted_consts)
+
+  def converted_fun(*args_hconsts):
+    num_args = len(args_hconsts) - num_consts
+    args, hoisted_consts = split_list(args_hconsts, [num_args])
+    consts = merge(closure_consts, hoisted_consts)
+    all_args, in_tree2 = tree_flatten(tuple(args))
+    assert in_tree == in_tree2
+    out_flat = core.eval_jaxpr(jaxpr, consts, *all_args)
+    return tree_unflatten(out_tree, out_flat)
+
+  return converted_fun, hoisted_consts
+
+def partition_list(choice, lst):
+  out = [], []
+  which = [out[choice(elt)].append(elt) or choice(elt) for elt in lst]
+  def merge(l1, l2):
+    i1, i2 = iter(l1), iter(l2)
+    return [next(i2 if snd else i1) for snd in which]
+  return out, merge
+
+def abstractify(x):
+  return core.raise_to_shaped(core.get_aval(x))
