+++ jax/experimental/ode.py
@@ -32,60 +32,16 @@ import operator as op
 import jax
 import jax.numpy as jnp
 from jax import core
-from jax import dtypes
+from jax import custom_derivatives
 from jax import lax
-from jax.util import safe_map, safe_zip, cache, split_list
-from jax.api_util import flatten_fun_nokwargs
+from jax.util import safe_map, safe_zip
 from jax.flatten_util import ravel_pytree
-from jax.tree_util import tree_map, tree_flatten, tree_unflatten
-from jax.interpreters import partial_eval as pe
+from jax.tree_util import tree_map
 from jax import linear_util as lu
-from jax import config
 
 map = safe_map
 zip = safe_zip
 
-
-@cache()
-def closure_convert(fun, in_tree, in_avals):
-  if config.omnistaging_enabled:
-    wrapped_fun, out_tree = flatten_fun_nokwargs(lu.wrap_init(fun), in_tree)
-    jaxpr, out_pvals, consts = pe.trace_to_jaxpr_dynamic(wrapped_fun, in_avals)
-  else:
-    in_pvals = [pe.PartialVal.unknown(aval) for aval in in_avals]
-    wrapped_fun, out_tree = flatten_fun_nokwargs(lu.wrap_init(fun), in_tree)
-    with core.initial_style_staging():  # type: ignore
-      jaxpr, out_pvals, consts = pe.trace_to_jaxpr(
-        wrapped_fun, in_pvals, instantiate=True, stage_out=False)  # type: ignore
-  out_tree = out_tree()
-
-  # We only want to closure convert for constants with respect to which we're
-  # differentiating. As a proxy for that, we hoist consts with float dtype.
-  # TODO(mattjj): revise this approach
-  is_float = lambda c: dtypes.issubdtype(dtypes.dtype(c), jnp.inexact)
-  (closure_consts, hoisted_consts), merge = partition_list(is_float, consts)
-  num_consts = len(hoisted_consts)
-
-  def converted_fun(y, t, *hconsts_args):
-    hoisted_consts, args = split_list(hconsts_args, [num_consts])
-    consts = merge(closure_consts, hoisted_consts)
-    all_args, in_tree2 = tree_flatten((y, t, *args))
-    assert in_tree == in_tree2
-    out_flat = core.eval_jaxpr(jaxpr, consts, *all_args)
-    return tree_unflatten(out_tree, out_flat)
-
-  return converted_fun, hoisted_consts
-
-def partition_list(choice, lst):
-  out = [], []
-  which = [out[choice(elt)].append(elt) or choice(elt) for elt in lst]
-  def merge(l1, l2):
-    i1, i2 = iter(l1), iter(l2)
-    return [next(i2 if snd else i1) for snd in which]
-  return out, merge
-
-def abstractify(x):
-  return core.raise_to_shaped(core.get_aval(x))
 
 def ravel_first_arg(f, unravel):
   return ravel_first_arg_(lu.wrap_init(f), unravel).call_wrapped
@@ -213,11 +169,9 @@ def odeint(func, y0, t, *args, rtol=1.4e-8, atol=1.4e-8, mxstep=jnp.inf):
              "\n{}.")
       raise TypeError(msg.format(arg))
 
-  flat_args, in_tree = tree_flatten((y0, t[0], *args))
-  in_avals = tuple(map(abstractify, flat_args))
-  converted, consts = closure_convert(func, in_tree, in_avals)
-
-  return _odeint_wrapper(converted, rtol, atol, mxstep, y0, t, *consts, *args)
+  converted, consts = custom_derivatives.closure_convert(
+      func, (y0, t[0], *args))
+  return _odeint_wrapper(converted, rtol, atol, mxstep, y0, t, *args, *consts)
 
 @partial(jax.jit, static_argnums=(0, 1, 2, 3))
 def _odeint_wrapper(func, rtol, atol, mxstep, y0, ts, *args):
